Create Code for Microcontroller to control all components:
At the outset, the ESP32 is initialized, and the necessary pins for sensor connections 
are configured. This includes defining the pins for the DHT11 sensor (temperature 
and humidity), ultrasonic sensor (trigger and echo), and any other sensors employed 
in the system. 
The microcontroller continuously reads data from the connected sensors, including 
the DHT11 sensor for temperature and humidity and the ultrasonic sensor for 
distance. The acquired data is then used to make informed decisions regarding 
irrigation management. 
To enable remote control and monitoring, the ESP32 communicates with the mobile 
application. This involves setting up a communication protocol, such as Wi-Fi, and 
defining methods for receiving commands from the application and sending sensor 
data back.